pyproject.toml
[build-system]
requires = ["uv_build>=0.9.9,<0.10.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-name = "e2d_core"
namespace = true

[project]
name = "casparianflow"
version = "0.1.0"
description = "..."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pandas>=2.2.2",
    "pydantic-settings>=2.12.0",
    "sqlalchemy>=2.0.44",
    "numpy>=1.26.0",
    "pyodbc>=5.0.0",
    "tqdm>=4.67.1",
    "pytest>=9.0.1",
    "pyarrow>=22.0.0",
    "pyzmq>=27.1.0",
    "openai>=2.12.0",
    "anthropic>=0.75.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.1",
    "ipykernel>=7.1.0",
    "pre-commit>=4.4.0",
    "ruff>=0.14.4",
    "pygount>=3.1.0",
]

[tool.ruff]
# Extend this list for any directories that are not part of your
# source distribution (e.g., deployment scripts, notebooks, tests).
exclude = [
    "scripts/",
    "tests/",
    "venv/",
    "__pypackages__/",
]
